+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- [BREAKING] Requires NodeJS 16+.
+### Deleted
+- Remove `IpregistryConfigBuilder.withApiUrl`. Use `IpregistryConfigBuilder.withBaseUrl` instead.
 
 ## [4.5.0] - 2022-04-15
 ### Added